Evaluate the best combination of parameters

Evaluate parameters that best fit for 'GreenFeed' data

Usage

eval_param(data, start_date, end_date, cutoff)

Arguments

data

a data frame with preliminary or finalized 'GreenFeed' data

start_date

a character string representing the start date of the study (format: "mm/dd/yyyy")

end_date

a character string representing the end date of the study (format: "mm/dd/yyyy")

cutoff

an integer specifying the range for identifying outliers (default: 3 SD)

Value

A data frame with the mean, SD, and CV for gas production using all possible combination of parameters

Examples

file <- readr::read_csv(system.file("extdata", "StudyName_GFdata.csv", package = "greenfeedr"))
#> Rows: 256 Columns: 21
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (6): AnimalName, RFID, StartTime, EndTime, InterruptingTags, RunTime
#> dbl  (13): FeederID, CO2GramsPerDay, CH4GramsPerDay, O2GramsPerDay, H2GramsP...
#> lgl   (1): WasInterrupted
#> time  (1): GoodDataDuration
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
eval <- eval_param(data = file,
                   start_date = "2024-05-13",
                   end_date = "2024-05-20"
                  )
#> All parameter combinations are being evaluated...
#> Done!