Download Daily GreenFeed Data via API

`get_gfdata()` downloads from C-Lock server daily GreenFeed data via an application programming interface (API). This function retrieves data based on specified parameters such as login information, date range and units, providing an Excel file with GreenFeed data.

Usage

get_gfdata(
  user,
  pass,
  exp = NA,
  unit,
  start_date,
  end_date = Sys.Date(),
  save_dir = getwd()
)

Arguments

user

User name to log in to the GreenFeed system

pass

Password to log in to the GreenFeed system

exp

Study name or other study identifier. It is used as the file name to save the data

unit

GreenFeed unit number(s). If multiple units, they could be in a vector, list, or character as "1,2"

start_date

Start date of the study

end_date

End date of the study. By default the current date is used

save_dir

Directory to save the output file. By default the current working directory is used

Value

This function does not return any value. It saves the downloaded data as a CSV file in the specified directory.

Examples

if (FALSE) { # has_credentials()
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")
exp <- "StudyName"
start_date <- "2023-01-01"
end_date <- Sys.Date()
save_dir <- tempdir()

# Example with multiple units as a comma-separated string
unit <- "304,305"
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with a single unit as a numeric value
unit <- 304
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with units as a vector
unit <- c(304, 305)
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)
}