Process GreenFeed Pellet Intakes

Processes the feedtimes file from GreenFeed system, including visits and food drops across a specific period, and it is used to calculate pellet intakes per animal from all units. Aggregates data to provide insights into the feeding behavior and pellet consumption of the animals during a study.

Arguments

file_path

File with feedtimes from C-Lock. If multiple files are provided, units should be in the same order

unit

GreenFeed unit number(s). If multiple units, they could be in a vector, list, or character as "1,2"

gcup

Grams of pellets per cup. By default 34 grams, but users should provide the grams based on the 10-drops test

start_date

Start date of the study

end_date

End date of the study

save_dir

Directory where to save the resulting file with pellet intakes

rfid_file

The file that contains the rfid of the animals enrolled in the study. The order should be col1=FarmName and col2=RFID

Value

An Excel file with pellet intakes for all animals and days within the specified period is saved to `save_dir`. The file is named "Pellet_Intakes_YYYY-MM-DD_YYYY-MM-DD.csv".

Examples

# You should provide the 'feedtimes' file provided by C-Lock.
# it could be a list of files if you have data from multiple units to combine
file <- list(system.file("extdata", "feedtimes.csv", package = "greenfeedr"))

# By default the function use 34g, but you should include the result obtained from the 10-drops test

# If the user include an rfid file, the structure should be in col1 the farmname or visualID, and
# col2 the RFID or TAG_ID. The file could be save in different formats (.xlsx, .csv, or .txt).

pellin(file,
  unit = 1,
  gcup = 34,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  save_dir = tempdir()
)
#> Please set the 'gcup' parameter based on the 10-drops test. The default value is 34g.
#> RFID is NULL. It is recommended to include it.
#> Pellet intakes file created and saved to /tmp/RtmpOT62S6