Download and Report GreenFeed Data

Generates PDF reports of daily and final GreenFeed data. If option "daily" is used, data is retrieved from C-Lock server via an application programming interface (API) and generates a PDF report to with number of animals, records, and gas production of the ongoing study. But, if option "final" is used, final data should be provided to generates a PDF report to evaluate all GreenFeed data obtained from the finalized study.

Arguments

user

User name to log in to the GreenFeed system

pass

Password to log in to the GreenFeed system

exp

Study name or other study identifier. It is used as the file name to report the data

unit

GreenFeed unit number(s). If multiple units, they could be in a vector, list, or character as "1,2"

start_date

Start date of the study

end_date

End date of the study. By default the current date is used

input_type

Input data could be from daily or final report: daily or final

save_dir

Directory to save the output file. By default the current working directory is used

plot_opt

Type of gas to plot: All, or CH4, CO2, O2, H2. By default only CH4 will be processed and reported

rfid_file

File that contains RFID of the animals in the study

file_path

List of files with final report from GreenFeed

Value

A CSV file with daily GreenFeed data and a PDF report with a description of the daily or final records.

Examples

if (FALSE) { # has_credentials()
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")

# The data range must be fewer than 180 days
# Example without rfid_file (by default NA)

report_gfdata(user,
  pass,
  exp = "StudyName",
  unit = 1,
  start_date = "2023-01-01",
  end_date = Sys.Date(),
  input_type = "daily",
  save_dir = tempdir(),
  plot_opt = "All"
)
}
# Create a Final Report in PDF format from the finalized data received from C-Lock Inc.
# Note that Unit could be numeric or character (It will use to print in the PDF)

# File is a list of reports from C-Lock inc.
# it could be one or multiples depend on the number of units
file <- list(system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr"))

# By default `report_gfdata()` plot only methane (CH4), but here we defined to plot "All" gases

# Is it possible to include a file with Farm IDs to use in the reports
# The file structure should be: FarmName | RFID

report_gfdata(
  user = NA,
  pass = NA,
  exp = "StudyName",
  unit = 1,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  input_type = "final",
  save_dir = tempdir(),
  plot_opt = "All",
  rfid_file = NA,
  file_path = file
)
#> No valid data provided. Please include a valid 'rfid_file' parameter.
#> 
#> 
#> processing file: FinalReportsGF.Rmd
#> 1/8                              
#> 2/8 [libraries]                  
#> 3/8                              
#> 4/8 [Data per cow]               
#> 5/8                              
#> 6/8 [Data per day]               
#> 7/8                              
#> 8/8 [Gas data production per cow]
#> output file: FinalReportsGF.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS FinalReportsGF.knit.md --to latex --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpJK5fc2/FinalReport_StudyName.tex --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --highlight-style tango --pdf-engine pdflatex --variable graphics --variable 'geometry:margin=1in' 
#> Warning: error in running command
#> ! sh: 1: pdflatex: not found
#> Error: LaTeX failed to compile /tmp/RtmpJK5fc2/FinalReport_StudyName.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See FinalReport_StudyName.log for more info.